% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{sfr_dbi_connection}
\alias{sfr_dbi_connection}
\title{Get an skiLift DBI connection from an sfr_connection}
\usage{
sfr_dbi_connection(conn)
}
\arguments{
\item{conn}{An \code{sfr_connection} object from \code{\link[=sfr_connect]{sfr_connect()}}.}
}
\value{
An \code{skiLift::SnowflakeConnection} object.
}
\description{
Returns an \code{skiLift::SnowflakeConnection} that can be used with
standard DBI methods (\code{dbGetQuery}, \code{dbWriteTable}, etc.) and dbplyr.
The connection is created lazily on first call and cached on the
\code{sfr_connection} object.
}
\details{
Requires the \code{skiLift} package to be installed.
}
\examples{
\dontrun{
sfr_conn <- sfr_connect(name = "my_profile")
dbi_con  <- sfr_dbi_connection(sfr_conn)
DBI::dbGetQuery(dbi_con, "SELECT 1")
}

}
